Revamp the process launching model
Add a concept to the implementation of a process handle, and add an
impl.spawn() to generate these.  main() will no longer spawn the first
process, that's just passed down for the lua layer to handle.  The lua
layer doesn't see too many major changes from this just yet, we're simply
spawning the process and storing its handle in the orch-global `process`
and using that.

In the future, we may break out a spawn command to orch scripts and allow
match actions to specify a process to match against, or something to that
effect.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>