Improve diagnostics on match failure
Provide at least the filename + line number, pattern, so that we know where
to even begin debugging failures.  If the match is parameterized then this
could be a little less helpful, but one can always stick a debug() call just
before match in question to log more information.

If all one() patterns fail, we do not spit out each pattern individually, we
just note that the overall one() did not have a successful match.

eof() will also produce a diagnostic on failure.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>