Add support for POSIX pattern matching
When the match object is created, we'll regcomp() it and get the pattern
object back from orchlua.  We can then subsequently :find() on the result
without it tracking any state, only operating on what we feed it.

Internally, pattern matchers may now have a compile() function that returns
an object.  The resulting object will be passed to the matcher instead of
the pattern itself.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>