examples: add a silly nc example, polish and install these
The nc example demonstrates that we can be executable (note the shebang),
and it demonstrates spawning an application in your orch script.

Rename the cat example to match current convention of naming these .orch,
and rewrite the introductory text.  Namely, we don't *need* to pipe scripts
in via stdin anymore (and the instructions to do so were not updated to use
the `-f` option like is needed as of late).

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>