Port orch(1) to OpenBSD
We'll just build orch_compat.c all the time now; any potentially cnoflicting
symbols are already preproc-gated behind OS that need it.

OpenBSD's posix_openpt doesn't support setting O_CLOEXEC, so set it after
the fact.

All tests pass, with exception to the shebang test.  OpenBSD's env(1) does
not support -S splitting that we need in order for the script to be passed
in as the -f argument.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>