Add a `make check` and some related infrastructure
Allow ORCHLUA_PATH to be overridden in the environment, so that we can test
the orch.lua directly out of the repository.  Flip our preference of orch
binary to use in the basic test, we want to prefer testing with the binary
we just built.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>
(cherry picked from commit 1102dd6071d13b493251e4450106d335cff1c013)