lib: add tty functionality
This adds functionality to allocate a term and manipulate it.  Manipulation
is done at the .orch script level with an stty() function, that takes the
field to manipulate (cflag, iflag, lflag, oflag, cc), a set field, and an
unset field.  Valid values for each of these are available in the tty table
that is exported to the .orch script.

Fixes #3

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>