Add an IPC registration mechanism, use it for child errors
There are a number of failure points where we might notice the child has
exited prematurely, but we have no clue why.  Have the parent register an
error handler prior to spawning a new child, and print those out if they
come in.

Have all of the error paths in the child fire off an error message before
they exit.

Messages are popped in orch_ipc_close() so that handlers still get called
when we're closing the process.  We now keep track of the lua_State in the
orch_process data in case we eventually want the error handler, or perhaps
another handler, to call back into orch.lua.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>