Start exporting an interface from orch.lua
We'll begin with run_script(); stop calling it directly, and instead add
that to the table we'll export.  Call it from the orch(1) driver rather
than passing scriptf as a parameter to orch_lua.c bits.

Nothing used impl.script or impl.script_root, so we'll just banish those off
for now and let them be orch.lua state instead, if it wants to keep them.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>