Push initial spawn() off into orch.run_script()
We'll start taking a config table for run_script(), and if config.command is
set we'll execute that.

Make the necessary change on the C side.  This reduces orchlua_cfg to just
a dirfd, which I've left in-place for now... this might get redesigned,
though I still have some vague idea of being able to include in an orch
script and only being able to include scripts in the same tree.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>