lib: make ipc slightly less fragile
Each process now gets an IPC handle that takes ownership of the ipc socket
and tracks the queue and any associated callbacks.  This gets passed around
to make sure that we're always operating on the right socket and the right
received data, rather than potentially crossing wires if some packets were
still in the queue from a process that's since disappeared.

This fixes an issue seen in the tty tests where we end up closing the socket
out from underneath the next process because another process has been
spawned before our current process is GC'd.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>