lib: clear up some valgrind complaints
Valgrind seems to see some false positives in what's initialized, and to be
fair it's a little tricky looking.  For ipc messages on the stack, just use
C99 initializers, and for heap-allocated messages just zero them out to be
safe.

In general we shouldn't be trying to close(-1), so add some assertions in
places that make sense to avoid it; triggering them may mean we're doing
something we shouldn't be.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>