lib: designate tag 0 as no-xmit
We'll bail out if we see this come over the wire as something went wrong.
It will be used later to signal a tag that is deallocated, as we may stash
a couple of zero-payload tags in the orch_ipc handle and push all of the
details of the orch_ipc_msg into orch_ipc.c.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>