lib: just make orch_ipc->callbacks an array
This was originally made a pointer because I thought that I might want to
do lazy-allocation for it, but it occurred to me that both sides will
generally want to install some callbacks at some point.  It turns out that's
still correct, so just allocate callback slots up front instead of managing
them separately.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>