lib: push the process into a wrapper table
Proxy through the methods available on our userdata to make this a little
more transparent.  The most immediate effect is that we can cleanly attach
our terminal state to the process; next we'll move the match buffer out to
the process wrapper since it doesn't really need to be the process' only
uservalue.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>