lib: move the MatchBuffer and Process out into orch.process
The global `process` in orch.lua is given a more accurate name; it's the
current process being acted on.  The Process and MatchBuffer now take an
execute function, because we may execute callbacks differently depending on
the context in which a process is spawned.  For instance, if we break
spawn() out to other lua libs, then those processes may just pcall any
callbacks they encounter -- there's no concept of a queue there.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>