tests: add a basic test of the 'direct' mode via the lib
libtest is currently just a custom package loader setup so that we can just
test porch libs right out of the build tree, rather than requiring them to
be installed first.

These tests are meant to be relatively primitive; no test framework, just a
series of assertions that we'll use to fail the check.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>