tests: add a check-install target
This one is functionally equivalent to both check-cli and check-lib, but
instead only operates on whatever's installed on the system.  This is useful
for smoke-testing the install layout and is expected to be used by, e.g.,
packaging systems.

The check-cli and check-lib targets will still be used for the aggregate
`make check` target, so that we don't need to count on being able to install
porch(1) onto the system to test what we just built.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>