lib: drain the buffer before close to fix on macOS
macOS will explicitly wait[0] for the controlling tty to be drained of
output when the controlling process exits.  This means that if the script
does not consume *all* of the output or wait for EOF prior to finishing, we
end up hanging at process exit for an indefinite amount of time.

If the process finished naturally before we tried to SIGINT it, then
delivering the signal is harmless even if it wouldn't have caught the signal
since it's already in the process of exiting at the point of that ttywait().

We apply this wait unconditionally for all platforms to provide both
stronger and consistent guarantees about the state of the output buffer
after process close.  Notably, we want to guarantee that we collected
anything that would have been output before we tried to SIGINT the process.

[0] https://github.com/apple-oss-distributions/xnu/blob/rel/xnu-10063/bsd/kern/kern_exit.c#L2311

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>