test: minor fixes to libtest.lua
Normalize PORCHLUA_PATH and PORCHLIB_PATH if they're set but empty.  If we
don't have either one set, then we don't need to install the package
searcher in the first place.  Allow only one of them to be set in the
searcher, though, and fallthrough to the other searchers as appropriate.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>