lib: don't try to drain the buffer on an unreleased process
Depending on where we're at in the closing of the process, we could easily
catch a SIGPIPE here if we, e.g., sent a SIGINT and killed the process.

Coincidentally, stop spawning a cat(1) process that we won't be using for
the resize tests.  The implicit close from immediately spawning off the
shell script tripped over this drain error intermittently.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>