ci: start linting manpages, too
Add a check-lint target that the check target depends on, and run our custom
lint-man.sh script as part of that.  This script will check both mandoc and
igor when installed and escalate any linting errors.

Start installing igor on our FreeBSD image, and bump versions to recently
supported FreeBSD.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>