lib: avoid fatal termination as much as possible
It's more idomatic to return nil, error string and let the caller decide how
they want to proceed than it is to terminate the process.  The lib state can
already be reasonably reset for direct consumers, but there's really no path
to recovery for porch(1) users.

pcall the processing in our state driver so that internally we can continue
to use error() to jump up the stack a bit, while still allowing the caller
to avoid termination.

Additionally, exit() has been augmented to avoid calling os.exit().
Instead, it calls error() with a table that indicates the exit code we're
supposed to use, which is then returned from run_script() for the caller to
decide what to do with.  `allow_exit` has been added as a configuration
option for the binary caller to request run_script() to just os.exit() as we
did before, just to simplify the C side.  run_script() now returns either
true to indicate that it succeeded, or an exit code.

Fixes #21

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>