src: add an rporch(1) command
This is used to execute an orch(5) script against a remote machine.  The
rsh program defaults to ssh, but can be swapped out at both the environemnt
and command line argument level.

Note that the script execution is still entirely local, it just executes the
spawned process over ssh.

Fixes #24

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>