lib: implement an .orch script generator
This uses the newly-introduced proxy mode to pass user input to the process,
and passes process output back up to stdout.  When user input is accepted,
the generator will generate a match statement for the last line of input
with some comments for previous lines in case we got it wrong.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>