lib: core: add a new process status type to capture wait(2) status
exit/signal/stop and their associated codes or signals are all I envision us
needing in the future.  Since extracting any of this information from the
status is pretty quick and the status itself is only a 32-bit int, we'll
just use a simple `int *` for our userdata.

Note that nothing synthesizes these just yet, but in the next commit we will
have process:eof() generate one.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>