lib: core: plumb a timeout through to process:eof()
process:eof() serves two purposes: to check for EOF, and to relay the exit
status of the process.  Before this commit, we'd just hang indefinitely
until the process closes if we did see EOF, which is perhaps not all that
useful.  Now, we'll wait up to the timeout specified.  For timeout=0, we
just set WNOHANG to do a quick probe and see if the process has already
terminated.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>