tests: add some framework for skipping lua tests on certain platforms
We're going to skip test_eof_timeout.lua entirely on Darwin for now, as we
seem to be hitting some weird edge-case in XNU that I don't have time to
diagnose at the moment.  Bash verifiably closes stdin/stdout/stderr, but
select(2) doesn't return for us to read EOF until the process exits.  This
breaks our expectations for our :eof(0) test, which is trying to ensure that
we can get back eof == true without the process having exited.

The test works on both FreeBSD and Ubuntu, so we're just disabling it on
Darwin for now.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>