share: lua: add an exec() action
This does another io.popen(), but instead of piping it to a spawned command
it instead pipes it to a line collector (if specified).  Whether the caller
is collecting output or not, we'll call a termfn if it's specified with the
wait status -- this is expected to be used for, e.g., things like FreeBSD's
mailx(1) signal tests where it may want to do a `test -f dead.letter` at
the end and pass/fail based on whether it exists.

Note that it's a little harder to construct pipelines where we do things
like `mktemp` and then use the result, since everything in the script is
queued.  I expect less of that type of usage, and perhaps mktemp is worth
having its own builtin functionality.

exec() is limited to the scripted environment; the direct usage has a full
lua available and can manage this themselves.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>