lib: core: record the signal mask when we spawn
We won't assume that the process signal mask has *not* changed because we
could be used in direct mode where it's feasible for some other resident of
the process to alter it.  We'll use this next to do signal mask alterations
in the parent then send them over to the child to apply.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>