lib: core: GC dead code and fix more portability issues
sigismember() doesn't return -1 for errors on all platforms; notably, it's
a simple boolean on macOS at least.  On those kinds of platforms, we'll let
the compiler figure it out and make sure we're not using sigismember() type
probing to determine the max signal.  After further examination, it looks
like FreeBSD is the only platform at the moment where we can't trust NSIG
for some signals that one could conceivably want to use.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>