share: lua: change sigunignore's primary name to 'sigcatch'
The underlying Process method isn't called 'sigcatch' because the C method
to fetch/set this stuff is called 'sigcatch', which makes sense.  'sigblock'
does not have this problem because it's an action we do to the signal mask,
which is normal verbiage.

This is probably fine as-is, but I don't feel like changing the underlying
terminology to accommodate when the user won't know the difference; they
only get exposed to the action (directly or via DirectProcess).  We'll just
document sigignore/sigcatch, sigblock/sigunblock as the logical actions and
leave sigunignore as an alias in case someone goes lib spelunking and
decides it might work.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>