tests: add basic stop/continue tests
This (stop)watch program is not a traditional stopwatch, but rather a
watcher for being in a stop state.  If we observe the application having
been suspended for at least as long as the caller specifies, we exit.  This
test is rather silly, but it gets the point across that we're actually doing
suspend properly.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>