tests: unbreak the stop/continue tests in my NetBSD VM
In this VM, we consistently manage to SIGSTOP the stopwatch before it even
manages to record the start time.  I had placed this intentionally just to
make sure that we didn't have any unwanted latency, but indeed that does
pose a reliability problem -- move it to later and assume that we won't
see 3+s worth of latency in trying to write to stdout.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>