lib: core: implement the C side of setid()
Provide a generic IPC message for setting either or both of the UID and GID
of the child process, and plumb it out through cprocess:setid(uid, gid).

nil or none for uid or gid will leave that particular attribute untouched
in the child process, and the active uid + gid are returned.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>