share: lua: expose gid/uid/setid actions for uid/gid functionality
uid/gid are special cases as they reflect the uid/gid of porch itself,
rather than the process spawned.  For the scripted mode, there is no way
to fetch the current credentials of the process -- the user just has to know
if they had called setid() and it did not error out, then the current user
has been changed.

The above note only applies to scripted mode; in direct mode, we expose them
on the DirectProcess and they just do what one would expect.

setid() does as advertised and sets the user, the group, or both of a
spawned process.  porch(1) itself will not drop privileges, but it will drop
the privileges of the spawned process.  The user should avoid calling setid
if it doesn't *know* that it can drop privileges.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>