porch, rporch: add -i include functionality
Any file specified via -i should be a lua script that will be evaluated via
dofile() after the scripted sandbox environment is setup, but before the
script has been evaluated.  The include file is expected to return a table
whose members will be added to the script's global environment, expanding
the sandbox.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>