share: lua: allow write() input to be suppressed in logs
It may be the case that logging our input to the process results in
noise that isn't needed, or in some cases it could be a security issue.

Add a parameter to suppress writes; in the near future, I think I'll
add a per-write option to suppress the log message or to specify a
replacement as a 'redaction' mechanism.

Signed-off-by: Kyle Evans <kevans@FreeBSD.org>